+++ src/tightdb/group_shared.cpp
@@ -576,12 +576,14 @@ void SharedGroup::open(const string& path, bool no_create_file,
 {
     TIGHTDB_ASSERT(!is_attached());
 
-    m_file_path = path + ".lock";
+    m_db_path = path;
+    m_key = key;
+    m_lockfile_path = path + ".lock";
     SlabAlloc& alloc = m_group.m_alloc;
 
     while (1) {
 
-        m_file.open(m_file_path, File::access_ReadWrite, File::create_Auto, 0);
+        m_file.open(m_lockfile_path, File::access_ReadWrite, File::create_Auto, 0);
         File::CloseGuard fcg(m_file);
         if (m_file.try_lock_exclusive()) {
 
@@ -813,6 +815,51 @@ void SharedGroup::open(const string& path, bool no_create_file,
 #endif
 }
 
+bool SharedGroup::compact()
+{
+    // Verify that preconditions for compacting is met:
+    if (m_transact_stage != transact_Ready) {
+        throw runtime_error(m_db_path + ": compact is not supported whithin a transaction");
+    }
+    string tmp_path = m_db_path + ".tmp";
+    SharedInfo* info = m_file_map.get_addr();
+    RobustLockGuard lock(info->controlmutex, &recover_from_dead_write_transact); // Throws
+    if (info->num_participants > 1)
+        return false;
+
+    // Using begin_read here ensures that we have access to the latest and greatest entry
+    // in the ringbuffer. We need to have access to that later to update top_ref and file_size.
+    begin_read();
+
+    // Compact by writing a new file holding only live data, then renaming the new file
+    // so it becomes the database file, replacing the old one in the process.
+    m_group.write(tmp_path, m_key, info->latest_version_number);
+    rename(tmp_path.c_str(), m_db_path.c_str());
+    end_read();
+
+    // We must detach group complety to force it to fully refresh its accessors for use
+    // in later transactions
+    m_group.complete_detach();
+    SlabAlloc& alloc = m_group.m_alloc;
+
+    // close and reopen the database file.
+    alloc.detach();
+    bool skip_validate = false;
+    bool no_create = true;
+    bool read_only = false;
+    bool is_shared = true;
+    bool server_sync_mode = false;
+    ref_type top_ref = alloc.attach_file(m_db_path, is_shared, read_only, no_create, 
+                                         skip_validate, m_key, server_sync_mode); // Throws
+    size_t file_size = alloc.get_baseline();
+
+    // update the versioning info to match
+    Ringbuffer::ReadCount& rc = const_cast<Ringbuffer::ReadCount&>(info->readers.get_last());
+    TIGHTDB_ASSERT(rc.version == info->latest_version_number);
+    rc.filesize = file_size;
+    rc.current_top = top_ref;
+    return true;
+}
 
 uint_fast64_t SharedGroup::get_number_of_versions()
 {
@@ -835,8 +882,12 @@ void SharedGroup::open(Replication& repl, DurabilityLevel dlevel, const char* ke
 
 #endif
 
-
 SharedGroup::~SharedGroup() TIGHTDB_NOEXCEPT
+{
+    close();
+}
+
+void SharedGroup::close() TIGHTDB_NOEXCEPT
 {
     if (!is_attached())
         return;
@@ -864,10 +915,8 @@ SharedGroup::~SharedGroup() TIGHTDB_NOEXCEPT
             // we can delete it when done.
             if (info->flags == durability_MemOnly) {
                 try {
-                    size_t path_len = m_file_path.size()-5; // remove ".lock"
-                    string db_path = m_file_path.substr(0, path_len); // Throws
                     m_group.m_alloc.detach();
-                    util::File::remove(db_path.c_str());
+                    util::File::remove(m_db_path.c_str());
                 }
                 catch(...) {} // ignored on purpose.
             }