+++ release_notes.md
@@ -19,6 +19,7 @@
   Operators are provided so that VersionID's may be compared.
 * Creating distinct views on integer, datetime, bool and enum columns is now possible.
 * Add `Table::minimum_datetime()` and `Table::maximum_datetime()`.
+* Extending `Table::get_sorted_view()` to support multi-column sorting.
 
 -----------
 
@@ -60,6 +61,7 @@
 
 ### Enhancements:
 
+* Added 'compact' method to SharedGroup for better control of compaction of the database file.
 * The following constants were added: `Group::max_table_name_length`,
   `Table::max_column_name_length`, `Table::max_string_size`, and
   `Table::max_binary_size`.